Add more enum example - Add enum with use - restructure enum related files
